# entity/compliance-entity/openapi.yaml
openapi: 3.0.3
info:
  title: GRC Platform – Compliance Entity Service
  description: Internal entity service exposing all audit and risk data from the GRC Platform MySQL database. Authentication
    is handled by the Choreo API Gateway; this service is only reachable by other Choreo components within the same organisation.
  version: 1.1.0
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
...
  /risks/search:
    post:
      tags:
      - Risks
      summary: Search risks
      operationId: searchRisks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRisksRequest'
      responses:
        '200':
          description: Search risks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchRisksResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /risks/{id}:
    get:
      tags:
      - Risks
      summary: Get a risk by ID
      operationId: getRiskById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Get a risk by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
      - Risks
      summary: Update a risk
      operationId: updateRisk
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRiskRequest'
      responses:
        '200':
          description: Update a risk
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /risks:
    post:
      tags:
      - Risks
      summary: Create a risk
      operationId: createRisk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRiskRequest'
      responses:
        '201':
          description: Create a risk
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /risks/next-sequence-number:
    get:
      tags:
      - Risks
      summary: Preview the next risk sequence number for a source register
      description: A MANAGED_SERVICES register counts per customer, so customerId is required for it and rejected for any other register.
      operationId: nextSequenceNumber
      parameters:
      - name: sourceRegisterId
        in: query
        schema:
          type: integer
        required: true
      - name: customerId
        in: query
        schema:
          type: integer
        required: false
      responses:
        '200':
          description: Next sequence number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NextSequenceResponse'
        '400': *id002
        '500': *id001
  /risks/{id}/detail:
    get:
      tags:
      - Risks
      summary: Get the fully-composed risk detail page
      operationId: getRiskDetail
      parameters:
      - name: id
        in: path
        schema:
          type: integer
        required: true
      responses:
        '200':
          description: Risk detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskDetail'
        '400': *id002
        '404': &id003
          $ref: '#/components/responses/NotFound'
        '500': *id001
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          default: 20
          maximum: 100
        offset:
          type: integer
          default: 0
...
    RiskLookupRef:
      type: object
      description: A register-template lookup value as it appears on one risk.
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
          description: Customers only.
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
...
    RiskScore:
      type: object
      properties:
        id:
          type: integer
        likelihood:
          type: integer
        impact:
          type: integer
        riskRating:
          type: integer
        riskLevel:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
        colorCode:
          type: string
...
    RiskComplianceReference:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdOn:
          type: string
          format: date-time
        updatedOn:
          type: string
          format: date-time
...
    Risk:
      type: object
      properties:
        id: &id004
          type: integer
        riskCode: &id005
          type: string
        riskYear: &id006
          type: integer
        riskQuarter: &id007
          type: string
          enum:
          - Q1
          - Q2
          - Q3
          - Q4
        riskTitle: &id008
          type: string
        riskDescription: &id009
          type: string
          nullable: true
        sourceRegisterId: &id010
          type: integer
        sourceRegisterName: &id011
          type: string
        assignmentTeamId: &id012
          type: integer
        assignmentTeamName: &id013
          type: string
        assignerId: &id014
          type: integer
        assignerUuid: &id015
          type: string
          description: Asgardeo id of the risk assigner. Empty when the row predates the identity migration and has not been backfilled.
        ownerId: &id016
          type: integer
        ownerUuid: &id017
          type: string
          description: Asgardeo id of the risk owner. Empty when the row predates the identity migration and has not been backfilled.
        workflowStatus: &id018
          type: string
          enum:
          - DRAFT
          - PENDING_RISK_OWNER_APPROVAL
          - PENDING_MANAGEMENT_APPROVAL
          - PENDING_COMPLIANCE_REVIEW
          - IN_REMEDIATION
          - PENDING_OWNER_COMPLETION_APPROVAL
          - PENDING_COMPLIANCE_CLOSURE
          - PENDING_AMENDMENT
          - PENDING_REVISION
          - ESCALATED
          - CLOSED
          - CANCELLED
        treatmentStrategy: &id019
          type: string
          enum:
          - REMEDIATE
          - ACCEPT
          - TRANSFER
          - AVOID
          nullable: true
        grossScoreId: &id020
          type: integer
          nullable: true
        grossRiskLevel: &id021
          type: string
          nullable: true
        implementationDate: &id022
          type: string
          nullable: true
        reassessmentDate: &id023
          type: string
          nullable: true
        createdOn: &id024
          type: string
          format: date-time
        updatedOn: &id025
          type: string
          format: date-time
        riskIdentifiedDate: &id026
          type: string
          nullable: true
        identifiedByType: &id027
          type: string
          nullable: true
          enum:
          - EMPLOYEE
          - EXTERNAL_PERSON
          - TOOL
        identifiedByName: &id028
          type: string
          nullable: true
        impactDescription: &id029
          type: string
          nullable: true
        actionPlanId: &id030
          type: integer
          nullable: true
        progress: &id031
          type: string
          nullable: true
        complianceApprovalBy: &id032
          type: integer
          nullable: true
        complianceApprovalDate: &id033
          type: string
          nullable: true
        gitIssueUrl: &id034
          type: string
          nullable: true
        emailSubject: &id035
          type: string
          nullable: true
        remarks: &id036
          type: string
          nullable: true
        riskType: &id037
          type: string
          enum:
          - NEW
          - UPDATED
        rejectionComment: &id038
          type: string
          nullable: true
        rejectionStage: &id039
          type: string
          nullable: true
        ownerFirstApprovedAt: &id040
          type: string
          nullable: true
        createdBy: &id041
          type: string
        updatedBy: &id042
          type: string
        effectiveRiskLevel: &id043
          type: string
          nullable: true
        effectiveColorCode: &id044
          type: string
          nullable: true
        registerTemplate: &id045
          type: string
          description: The source register's template. Empty if the register row is missing.
        customerName: &id046
          type: string
          nullable: true
          description: MANAGED_SERVICES risks only.
        environments: &id047
          type: array
          items:
            type: string
            enum:
            - PRODUCTION
            - NON_PRODUCTION
            - DR
          description: MANAGED_SERVICES risks only, in PRODUCTION, NON_PRODUCTION, DR order; empty otherwise.
        platformNames: &id048
          type: array
          items:
            type: string
          description: AGGREGATED risks only, ordered by name; empty otherwise.
    SearchRisksRequest:
      type: object
      properties:
        scopeSourceRegisterIds:
          type: array
          items:
            type: integer
          description: >-
            Risks RAISED in these registers. ORed with scopeAssignmentTeamIds.
        scopeAssignmentTeamIds:
          type: array
          items:
            type: integer
          description: >-
            Risks ROUTED to these teams. Separate from scopeSourceRegisterIds
            because different roles scope by different dimensions of a risk — a
            Risk Owner of HR sees work routed to HR, not risks HR raised. Both
            empty means unrestricted.
        searchQuery:
          type: string
        workflowStatusKeys:
          type: array
          items:
            type: string
        sourceRegisterIds:
          type: array
          items:
            type: integer
        assignmentTeamIds:
          type: array
          items:
            type: integer
        riskYears:
          type: array
          items:
            type: integer
        riskQuarterKeys:
          type: array
          items:
            type: string
        pagination:
          $ref: '#/components/schemas/Pagination'
        customerIds:
          type: array
          items:
            type: integer
        environmentKeys:
          type: array
          items:
            type: string
            enum:
            - PRODUCTION
            - NON_PRODUCTION
            - DR
        platformIds:
          type: array
          items:
            type: integer
    SearchRisksResponse:
      type: object
      properties:
        risks:
          type: array
          items:
            $ref: '#/components/schemas/Risk'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    CreateRiskRequest:
      type: object
      properties:
        riskTitle:
          type: string
        riskDescription:
          type: string
          nullable: true
        sourceRegisterId:
          type: integer
        assignmentTeamId:
          type: integer
        assignerId:
          type: integer
        ownerId:
          type: integer
        riskYear:
          type: integer
        riskQuarter:
          type: string
          enum:
          - Q1
          - Q2
          - Q3
          - Q4
        likelihood:
          type: integer
        impact:
          type: integer
        treatmentStrategy:
          type: string
          enum:
          - REMEDIATE
          - ACCEPT
          - TRANSFER
          - AVOID
          nullable: true
        implementationDate:
          type: string
          nullable: true
        reassessmentDate:
          type: string
          nullable: true
        impactDescription:
          type: string
          nullable: true
        riskIdentifiedDate:
          type: string
          nullable: true
        identifiedByType:
          type: string
          enum:
          - EMPLOYEE
          - EXTERNAL_PERSON
          - TOOL
          nullable: true
        identifiedByName:
          type: string
          nullable: true
        gitIssueUrl:
          type: string
          nullable: true
        emailSubject:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        actionOwnerId:
          type: integer
          nullable: true
        actionPlanDescription:
          type: string
          nullable: true
        actionSteps:
          type: array
          items:
            $ref: '#/components/schemas/ActionStepInput'
        complianceReferenceIds:
          type: array
          items:
            type: integer
        progress:
          type: string
          nullable: true
        platformIds:
          type: array
          items:
            type: integer
          description: AGGREGATED registers only; at least one, each an ACTIVE platform.
        customerId:
          type: integer
          nullable: true
          description: MANAGED_SERVICES registers only (required there); an ACTIVE customer. Its code goes into the risk code.
        deploymentTypeId:
          type: integer
          nullable: true
          description: MANAGED_SERVICES registers only (required there); an ACTIVE deployment type.
        productIds:
          type: array
          items:
            type: integer
          description: MANAGED_SERVICES registers only; at least one, each an ACTIVE product.
        environments:
          type: array
          items:
            type: string
            enum:
            - PRODUCTION
            - NON_PRODUCTION
            - DR
          description: MANAGED_SERVICES registers only; at least one, no repeats.
        createdBy:
          type: string
      description: >-
        Which register-template fields are allowed and required depends on the
        source register's registerTemplate (RISK_MODULE_DESIGN.md §14); anything
        else is a 400. A MANAGED_SERVICES register also rejects
        complianceReferenceIds and only accepts MANAGED_SERVICES assignment
        teams, while other registers reject those teams.
    UpdateRiskRequest:
      type: object
      properties:
        riskTitle:
          type: string
          nullable: true
        riskDescription:
          type: string
          nullable: true
        workflowStatus:
          type: string
          enum:
          - DRAFT
          - PENDING_RISK_OWNER_APPROVAL
          - PENDING_MANAGEMENT_APPROVAL
          - PENDING_COMPLIANCE_REVIEW
          - IN_REMEDIATION
          - PENDING_OWNER_COMPLETION_APPROVAL
          - PENDING_COMPLIANCE_CLOSURE
          - PENDING_AMENDMENT
          - PENDING_REVISION
          - ESCALATED
          - CLOSED
          - CANCELLED
          nullable: true
        treatmentStrategy:
          type: string
          enum:
          - REMEDIATE
          - ACCEPT
          - TRANSFER
          - AVOID
          nullable: true
        grossScoreId:
          type: integer
          nullable: true
        implementationDate:
          type: string
          nullable: true
        reassessmentDate:
          type: string
          nullable: true
        progress:
          type: string
          nullable: true
        rejectionComment:
          type: string
          nullable: true
        rejectionStage:
          type: string
          nullable: true
        complianceApprovalBy:
          type: integer
          nullable: true
        complianceApprovalDate:
          type: string
          nullable: true
        assignmentTeamId:
          type: integer
          nullable: true
        ownerId:
          type: integer
          nullable: true
        actionPlanId:
          type: integer
          nullable: true
        gitIssueUrl:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        emailSubject:
          type: string
          nullable: true
        riskType:
          type: string
          nullable: true
          enum:
          - NEW
          - UPDATED
        ownerFirstApprovedAt:
          type: string
          nullable: true
        impactDescription:
          type: string
          nullable: true
        riskIdentifiedDate:
          type: string
          nullable: true
        identifiedByType:
          type: string
          nullable: true
          enum:
          - EMPLOYEE
          - EXTERNAL_PERSON
          - TOOL
        identifiedByName:
          type: string
          nullable: true
        assignerId:
          type: integer
          nullable: true
        clearRejection:
          type: boolean
        complianceReferenceIds:
          type: array
          items:
            type: integer
        actionPlan:
          allOf:
          - $ref: '#/components/schemas/ActionPlanUpdate'
          nullable: true
        actionSteps:
          type: array
          items:
            $ref: '#/components/schemas/ActionStepUpdate'
        changeLog:
          type: array
          items:
            $ref: '#/components/schemas/ChangeLogEntry'
        updatedBy:
          type: string
        expectedStatus:
          type: string
        platformIds:
          type: array
          nullable: true
          items:
            type: integer
          description: AGGREGATED risks only. Null leaves platforms untouched; otherwise the complete set, at least one. Newly added platforms must be ACTIVE.
        deploymentTypeId:
          type: integer
          nullable: true
          description: MANAGED_SERVICES risks only; a new value must be ACTIVE. There is no customerId — a risk's customer never changes.
        productIds:
          type: array
          nullable: true
          items:
            type: integer
          description: MANAGED_SERVICES risks only. Same rules as platformIds.
        environments:
          type: array
          nullable: true
          items:
            type: string
            enum:
            - PRODUCTION
            - NON_PRODUCTION
            - DR
          description: MANAGED_SERVICES risks only. Null leaves them untouched; otherwise the complete set, at least one.
...
    RiskActionStep:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        stepNo:
          type: integer
        description:
          type: string
          nullable: true
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
        completedDate:
          type: string
          nullable: true
        createdOn:
          type: string
          format: date-time
        updatedOn:
          type: string
          format: date-time
...
    RiskAssessment:
      type: object
      properties:
        id:
          type: integer
        riskId:
          type: integer
        scoreId:
          type: integer
        residualLikelihood:
          type: integer
        residualImpact:
          type: integer
        residualRating:
          type: integer
        residualLevel:
          type: string
        residualColorCode:
          type: string
        progress:
          type: string
        reassessmentDate:
          type: string
        assessedBy:
          type: string
        createdOn:
          type: string
          format: date-time
...
    ActionStepInput:
      type: object
      properties:
        description:
          type: string
    ActionPlanUpdate:
      type: object
      properties:
        description:
          type: string
          nullable: true
        actionOwnerId:
          type: integer
          nullable: true
    ActionStepUpdate:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        description:
          type: string
    ChangeLogEntry:
      type: object
      properties:
        action:
          type: string
          enum:
          - CREATE
          - UPDATE
        fieldChanged:
          type: string
          nullable: true
        oldValue:
          type: string
          nullable: true
        newValue:
          type: string
          nullable: true
...
    NextSequenceResponse:
      type: object
      properties:
        nextSequenceNumber:
          type: integer
    RiskActionPlanDetail:
      type: object
      properties:
        id:
          type: integer
        riskId:
          type: integer
        actionOwnerId:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
        planType:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/RiskActionStep'
    RiskDetail:
      type: object
      description: 'The fully-composed risk returned by GET /risks/{id}/detail: every risk column plus resolved names, both
        scores, and related rows.'
      properties:
        id: *id004
        riskCode: *id005
        riskYear: *id006
        riskQuarter: *id007
        riskTitle: *id008
        riskDescription: *id009
        sourceRegisterId: *id010
        sourceRegisterName: *id011
        assignmentTeamId: *id012
        assignmentTeamName: *id013
        assignerId: *id014
        assignerUuid: *id015
        ownerId: *id016
        ownerUuid: *id017
        workflowStatus: *id018
        treatmentStrategy: *id019
        grossScoreId: *id020
        grossRiskLevel: *id021
        implementationDate: *id022
        reassessmentDate: *id023
        createdOn: *id024
        updatedOn: *id025
        riskIdentifiedDate: *id026
        identifiedByType: *id027
        identifiedByName: *id028
        impactDescription: *id029
        actionPlanId: *id030
        progress: *id031
        complianceApprovalBy: *id032
        complianceApprovalDate: *id033
        gitIssueUrl: *id034
        emailSubject: *id035
        remarks: *id036
        riskType: *id037
        rejectionComment: *id038
        rejectionStage: *id039
        ownerFirstApprovedAt: *id040
        createdBy: *id041
        updatedBy: *id042
        effectiveRiskLevel: *id043
        effectiveColorCode: *id044
        registerTemplate: *id045
        customerName: *id046
        environments: *id047
        platformNames: *id048
        complianceApproverName:
          type: string
          nullable: true
        grossScore:
          allOf:
          - $ref: '#/components/schemas/RiskScore'
          nullable: true
        effectiveScore:
          allOf:
          - $ref: '#/components/schemas/RiskScore'
          nullable: true
        complianceReferences:
          type: array
          items:
            $ref: '#/components/schemas/RiskComplianceReference'
        actionPlan:
          allOf:
          - $ref: '#/components/schemas/RiskActionPlanDetail'
          nullable: true
        assessments:
          type: array
          items:
            $ref: '#/components/schemas/RiskAssessment'
        customer:
          nullable: true
          allOf:
          - $ref: '#/components/schemas/RiskLookupRef'
        deploymentType:
          nullable: true
          allOf:
          - $ref: '#/components/schemas/RiskLookupRef'
        products:
          type: array
          items:
            $ref: '#/components/schemas/RiskLookupRef'
        platforms:
          type: array
          items:
            $ref: '#/components/schemas/RiskLookupRef'
...